+++ config/routes.rb
@@ -54,6 +54,9 @@ Rails.application.routes.draw do
 
   scope module: :blog do
     get :blog
+    get 'blog/category/:category', action: 'blog', as: :blog_category
+    get 'blog/author/:author', action: 'blog', as: :blog_author
+    get 'blog/:slug', action: 'show'
     get 'blog/:year/:month/:slug', action: 'show', as: :blog_post
   end